refactor(audio): split audio type tests for GH727
